// app/backend/src/Interfaces/LeaderBoardInterface.ts
export interface LeaderBoardInterface {
  name: string;
  totalPoints: number;
  totalGames: number;
  totalVictories: number;
  totalDraws: number;
  totalLosses: number;
  goalsFavor: number;
  goalsOwn: number;
  goalsBalance: number;
  efficiency: string;
}

export interface LeaderBoardModelInterface {
  getLeaderBoard(): Promise<LeaderBoardInterface[]>;
}